+++ dashboard/src/features/orgs/projects/database/common/utils/normalizeTableConstraints/index.ts
@@ -1,0 +1,1 @@
+export { default as normalizeTableConstraints } from './normalizeTableConstraints';
