+++ dashboard/src/features/orgs/projects/database/dataGrid/hooks/useTableQuery/fetchTable.ts
@@ -1,5 +1,6 @@
 import { formatWithArray } from 'node-pg-format';
 import { getPreparedReadOnlyHasuraQuery } from '@/features/orgs/projects/database/common/utils/hasuraQueryHelpers';
+import { normalizeTableConstraints } from '@/features/orgs/projects/database/common/utils/normalizeTableConstraints';
 import {
   COLUMN_DEFINITION_QUERY,
   CONSTRAINT_DEFINITION_QUERY,
@@ -16,7 +17,6 @@ import type {
   QueryResult,
   TableLikeObjectType,
 } from '@/features/orgs/projects/database/dataGrid/types/dataBrowser';
-import { extractForeignKeyRelation } from '@/features/orgs/projects/database/dataGrid/utils/extractForeignKeyRelation';
 import { POSTGRESQL_ERROR_CODES } from '@/features/orgs/projects/database/dataGrid/utils/postgresqlConstants';
 import { buildDefaultOrderByClause } from './buildDefaultOrderByClause';
 import { filtersToWhere } from './filtersToWhere';
@@ -216,77 +216,11 @@ export default async function fetchTable({
   const [, ...rawColumns] = responseData[0].result;
   const [, ...rawConstraints] = responseData[1].result;
 
-  const foreignKeyRelationMap = new Map<string, ForeignKeyRelation>();
-  const uniqueKeyConstraintMap = new Map<string, string[]>();
-  const primaryKeyConstraintMap = new Map<string, string[]>();
-
-  rawConstraints.forEach((rawConstraint) => {
-    const constraint = JSON.parse(rawConstraint);
-    const {
-      column_name: columnName,
-      constraint_type: constraintType,
-      constraint_name: constraintName,
-    } = constraint;
-
-    if (constraintType === 'f') {
-      const { constraint_definition: constraintDefinition } = constraint;
-      const foreignKeyRelation = extractForeignKeyRelation(
-        constraintName,
-        constraintDefinition,
-      );
-
-      // Composite foreign keys extract a combined columnName ("a, b") that
-      // never equals the constraint row's single column, so they are skipped.
-      if (
-        foreignKeyRelation &&
-        foreignKeyRelation.columnName === columnName &&
-        !foreignKeyRelationMap.has(columnName)
-      ) {
-        foreignKeyRelationMap.set(columnName, {
-          ...foreignKeyRelation,
-          referencedSchema: foreignKeyRelation.referencedSchema || schema,
-        });
-      }
-    }
-
-    if (constraintType === 'p') {
-      if (primaryKeyConstraintMap.has(columnName)) {
-        primaryKeyConstraintMap.set(columnName, [
-          ...primaryKeyConstraintMap.get(columnName)!,
-          constraintName,
-        ]);
-      } else {
-        primaryKeyConstraintMap.set(columnName, [constraintName]);
-      }
-    }
-
-    if (constraintType === 'u') {
-      if (uniqueKeyConstraintMap.has(columnName)) {
-        uniqueKeyConstraintMap.set(columnName, [
-          ...uniqueKeyConstraintMap.get(columnName)!,
-          constraintName,
-        ]);
-      } else {
-        uniqueKeyConstraintMap.set(columnName, [constraintName]);
-      }
-    }
-  });
-
-  const columns = rawColumns
-    .map((rawColumn) => {
-      const column = JSON.parse(rawColumn);
-      const foreignKeyRelation = foreignKeyRelationMap.get(column.column_name);
-
-      return {
-        ...column,
-        unique_constraints:
-          uniqueKeyConstraintMap.get(column.column_name) || [],
-        primary_constraints:
-          primaryKeyConstraintMap.get(column.column_name) || [],
-        foreign_key_relation: foreignKeyRelation ?? null,
-      } as NormalizedQueryDataRow;
-    })
-    .sort((a, b) => a.ordinal_position - b.ordinal_position);
+  const { columns, foreignKeyRelations } = normalizeTableConstraints(
+    rawColumns,
+    rawConstraints,
+    schema,
+  );
 
   if (!orderByClause) {
     orderByClause = buildDefaultOrderByClause(columns, tableType);
@@ -321,24 +255,6 @@ export default async function fetchTable({
     }),
   });
 
-  const flatForeignKeyRelations = Array.from(
-    foreignKeyRelationMap.values(),
-  ).reduce((accumulator, foreignKeyRelation) => {
-    const column = columns.find(
-      ({ column_name }) => column_name === foreignKeyRelation.columnName,
-    );
-
-    if (!column) {
-      return accumulator;
-    }
-
-    const foreignKeyWithOneToOne: ForeignKeyRelation = {
-      ...foreignKeyRelation,
-      oneToOne: column.is_unique || column.is_primary,
-    };
-    return [...accumulator, foreignKeyWithOneToOne];
-  }, [] as ForeignKeyRelation[]);
-
   const rawData: QueryResult<string[]> | QueryError =
     await rowDataResponse.json();
 
@@ -349,7 +265,7 @@ export default async function fetchTable({
       error:
         rawData.internal?.error.message ||
         'Something went wrong while fetching the table rows.',
-      foreignKeyRelations: flatForeignKeyRelations,
+      foreignKeyRelations,
       numberOfRows: 0,
     };
   }
@@ -361,7 +277,7 @@ export default async function fetchTable({
     columns,
     rows: rowData.map((row) => JSON.parse(row)) as NormalizedQueryDataRow[],
     error: null,
-    foreignKeyRelations: flatForeignKeyRelations,
+    foreignKeyRelations,
     numberOfRows: parseInt(rowAggregate, 10) || 0,
   };
 }
